Calculate undershoots for the mouse and myo.
-With Myo calculate an undershoot if the cursor has not yet reached the
rectangle but their velocity is 0.
-With the mouse, the mouseMoved() built in processing method can no longer be
used, because we actually need to detect when the mouse is *not* moving. So
in the draw loop check the current position of the mouse, and see if it is
greater than, less than or equal the mouse's previous position (mouseX and
pmouseX) (if they are equal, then we know the speed/velocity is 0).
-To calculate the undershoots we also have to make sure the cursor was moving
in the correct direction towards the next rectangle if they have stopped.
-We also check to make sure they have stopped on the correct side of the
rectangle.